Restore outline to focused links
Bitters removes this from focused links, instead relying on your
`:hover` state to be an adequate indicator of focus. This simply isn't
always true as I don't have any hover states on some footer links and
the hover state can sometimes be very subtle. There seems to be no
downsides to enabling the outline, so I've added it back.